requirements.txt
astroid==2.3.3
click==7.1.1
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
libsass==0.19.4
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
pyScss==1.3.7
python-dotenv==0.13.0
python-http-client==3.2.7
sendgrid==6.2.2
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
